remove chatroomstart container
